GeoJSON message bar in the itinerary view
